Revert "Added a few tests that measure the receiver’s rate."
This reverts commit 0c51959c9315f63bc80a7ff5b716f48f907b1152.
