+++ streaming/src/test/scala/org/apache/spark/streaming/ReceiverSuite.scala
@@ -155,6 +155,63 @@ class ReceiverSuite extends TestSuiteBase with Timeouts with Serializable {
     assert(recordedData.toSet === generatedData.toSet)
   }
 
+  ignore("block generator throttling") {
+    val blockGeneratorListener = new FakeBlockGeneratorListener
+    val blockIntervalMs = 100
+    val maxRate = 1001
+    val conf = new SparkConf().set("spark.streaming.blockInterval", s"${blockIntervalMs}ms").
+      set("spark.streaming.receiver.maxRate", maxRate.toString)
+    val blockGenerator = new BlockGenerator(blockGeneratorListener, 1, conf)
+    val expectedBlocks = 20
+    val waitTime = expectedBlocks * blockIntervalMs
+    val expectedMessages = maxRate * waitTime / 1000
+    val expectedMessagesPerBlock = maxRate * blockIntervalMs / 1000
+    val generatedData = new ArrayBuffer[Int]
+
+    // Generate blocks
+    val startTime = System.currentTimeMillis()
+    blockGenerator.start()
+    var count = 0
+    while(System.currentTimeMillis - startTime < waitTime) {
+      blockGenerator.addData(count)
+      generatedData += count
+      count += 1
+    }
+    blockGenerator.stop()
+
+    val recordedBlocks = blockGeneratorListener.arrayBuffers
+    val recordedData = recordedBlocks.flatten
+    assert(blockGeneratorListener.arrayBuffers.size > 0, "No blocks received")
+    assert(recordedData.toSet === generatedData.toSet, "Received data not same")
+
+    // recordedData size should be close to the expected rate; use an error margin proportional to
+    // the value, so that rate changes don't cause a brittle test
+    val minExpectedMessages = expectedMessages - 0.05 * expectedMessages
+    val maxExpectedMessages = expectedMessages + 0.05 * expectedMessages
+    val numMessages = recordedData.size
+    assert(
+      numMessages >= minExpectedMessages && numMessages <= maxExpectedMessages,
+      s"#records received = $numMessages, not between $minExpectedMessages and $maxExpectedMessages"
+    )
+
+    // XXX Checking every block would require an even distribution of messages across blocks,
+    // which throttling code does not control. Therefore, test against the average.
+    val minExpectedMessagesPerBlock = expectedMessagesPerBlock - 0.05 * expectedMessagesPerBlock
+    val maxExpectedMessagesPerBlock = expectedMessagesPerBlock + 0.05 * expectedMessagesPerBlock
+    val receivedBlockSizes = recordedBlocks.map { _.size }.mkString(",")
+
+    // the first and last block may be incomplete, so we slice them out
+    val validBlocks = recordedBlocks.drop(1).dropRight(1)
+    val averageBlockSize = validBlocks.map(block => block.size).sum / validBlocks.size
+
+    assert(
+      averageBlockSize >= minExpectedMessagesPerBlock &&
+        averageBlockSize <= maxExpectedMessagesPerBlock,
+      s"# records in received blocks = [$receivedBlockSizes], not between " +
+        s"$minExpectedMessagesPerBlock and $maxExpectedMessagesPerBlock, on average"
+    )
+  }
+
   /**
    * Test whether write ahead logs are generated by received,
    * and automatically cleaned up. The clean up must be aware of the
@@ -290,33 +347,28 @@ class ReceiverSuite extends TestSuiteBase with Timeouts with Serializable {
       errors += throwable
     }
   }
-}
 
-/**
- * An implementation of BlockGeneratorListener that is used to test the BlockGenerator.
- */
-class FakeBlockGeneratorListener(pushDelay: Long = 0) extends BlockGeneratorListener {
-  // buffer of data received as ArrayBuffers
-  val arrayBuffers = new ArrayBuffer[ArrayBuffer[Int]]
-  val errors = new ArrayBuffer[Throwable]
+  /**
+   * An implementation of BlockGeneratorListener that is used to test the BlockGenerator.
+   */
+  class FakeBlockGeneratorListener(pushDelay: Long = 0) extends BlockGeneratorListener {
+    // buffer of data received as ArrayBuffers
+    val arrayBuffers = new ArrayBuffer[ArrayBuffer[Int]]
+    val errors = new ArrayBuffer[Throwable]
 
-  def onAddData(data: Any, metadata: Any) {}
+    def onAddData(data: Any, metadata: Any) { }
 
-  def onGenerateBlock(blockId: StreamBlockId) {}
+    def onGenerateBlock(blockId: StreamBlockId) { }
 
-  def onPushBlock(blockId: StreamBlockId, arrayBuffer: ArrayBuffer[_]) {
-    val bufferOfInts = arrayBuffer.map(_.asInstanceOf[Int])
-    arrayBuffers += bufferOfInts
-    Thread.sleep(0)
-  }
+    def onPushBlock(blockId: StreamBlockId, arrayBuffer: ArrayBuffer[_]) {
+      val bufferOfInts = arrayBuffer.map(_.asInstanceOf[Int])
+      arrayBuffers += bufferOfInts
+      Thread.sleep(0)
+    }
 
-  def onError(message: String, throwable: Throwable) {
-    errors += throwable
-  }
-
-  def reset(): Unit = {
-    arrayBuffers.clear()
-    errors.clear()
+    def onError(message: String, throwable: Throwable) {
+      errors += throwable
+    }
   }
 }
